Require activesupport < 7.1 due to changes in ActiveSupport Broadcast inteface changes
In ActiveSupport < 7.1, the interface for broadcasting logs to multiple loggers overrides the add method of the logger to log to the broadcast logger.
This works with ContextualLogger because we only add new args to the debug, info, warn, and error methods.

In ActiveSupport 7.1, they changed the Broadcast Logger to now be its own separate class that takes a list of loggers and passes the args onto each logger. The problem I've found in AS 7.1 is if you setup the Broadcast logger with a ContextualLogger and a Ruby standard Logger, if you attempt to pass the context arg, this will fail with an invalid number of arguments on the standard ruby logger.